Add statistics, verbosity, fixed bit pattern options to random_integer_grids test.
